+++ src/main/java/project/forwork/api/common/infrastructure/enums/FieldType.java
@@ -1,4 +1,4 @@
-package project.forwork.api.domain.resume.infrastructure.enums;
+package project.forwork.api.common.infrastructure.enums;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import lombok.AllArgsConstructor;
