+++ src/main/java/project/forwork/api/domain/orderresume/controller/model/OrderResumeResponse.java
@@ -5,8 +5,6 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import project.forwork.api.domain.orderresume.infrastructure.enums.OrderResumeStatus;
-import project.forwork.api.domain.resume.infrastructure.enums.FieldType;
-import project.forwork.api.domain.resume.infrastructure.enums.LevelType;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
