+++ src/main/java/project/forwork/api/domain/salespost/infrastructure/enums/SalesPostSortType.java
@@ -1,5 +1,5 @@
 package project.forwork.api.domain.salespost.infrastructure.enums;
 
 public enum SalesPostSortType {
-    OLD, NEW, HIGHEST_PRICE, LOWEST_PRICE, VIEW_COUNT, BEST_SELLING
+    OLD, NEW, HIGHEST_PRICE, LOWEST_PRICE, VIEW_COUNT, BEST_SELLING, DEFAULT
 }
