+++ src/main/java/project/forwork/api/domain/salespost/service/port/SalesPostRepositoryCustom.java
@@ -1,19 +1,13 @@
 package project.forwork.api.domain.salespost.service.port;
 
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import project.forwork.api.common.infrastructure.enums.FieldType;
-import project.forwork.api.common.infrastructure.enums.LevelType;
+
 import project.forwork.api.domain.salespost.controller.model.SalesPostDetailResponse;
 import project.forwork.api.domain.salespost.controller.model.SalesPostFilterCond;
 import project.forwork.api.domain.salespost.controller.model.SalesPostResponse;
 import project.forwork.api.domain.salespost.controller.model.SalesPostSellerResponse;
-import project.forwork.api.domain.salespost.infrastructure.SalesPostSearchCond;
-import project.forwork.api.domain.salespost.infrastructure.enums.FieldCond;
-import project.forwork.api.domain.salespost.infrastructure.enums.LevelCond;
-import project.forwork.api.domain.salespost.infrastructure.enums.SalesPostSortType;
 
-import java.math.BigDecimal;
+
+
 import java.util.List;
 
 public interface SalesPostRepositoryCustom {
